requirements.txt
streamlit
crewai
PyPDF2
python-docx
